Generate THL color palette

Get the hex values of colors in predefined THL color palette. For possible values see the details.

palette_thl(name, n = NULL, force = FALSE)

Arguments

name

Spesify the name of the palette.

n

Number of colours to generate from palette.

force

Allows to generate more colors than the default palette has. Supressses the warnings produced from recycling or interpolating the colors.

Value

Character vector of the hex values of the colors

Details

Sequential, one-way palettes: “s1”, “s2”, “s3”, “s4”, “s5”, “s6”. Diverging, two-way palettes: “k1”, “k2”, “k3”, “k4”, “k5”, “k6”. Qualitative palettes:

“quali” or "laadullinen": all qualitative colors

“line” or "viiva": colors for lineplots

“area” or "pinta": colors for barplots or other plots with large areas to be filled

“theme” or "teema": other named colors

Examples

palette_thl("viiva", 6)
#> [1] "#519b2f" "#2f62ad" "#be3f72" "#29a0c1" "#faa61a" "#cc77ac"
palette_thl("line")
#> [1] "#519b2f" "#2f62ad" "#be3f72" "#29a0c1" "#faa61a" "#cc77ac" "#7bc143"