THL visual theme using ggplot2

Complete theme controlling all non-data display.

theme_thl(
  base_size = 16,
  base_family = "ArialMT",
  base_line_size = thlPtsConvert(1.5),
  base_rect_size = thlPtsConvert(1.5),
  horizontal = FALSE,
  show_axis_line = TRUE,
  ...
)

Arguments

base_size
base font size in PostScript points

base_family
base font family

base_line_size
base size for line elements

base_rect_size
base size for rect elements

horizontal
Set as TRUE if plotting horizontal bars.

show_axis_line
Should axis line be drawn? Axis line should be removed if the coordinate axes are not in the border of the plot

...
additional properties defined in theme()

Value

theme() object, to be combined with ggplot object with "+"

Examples

# add this to the usual plot with "+"
theme_thl()

#> List of 62
#>  $ line                 :List of 6
#>   ..$ colour       : chr "#dcdfe2"
#>   ..$ size         : num 0.529
#>   ..$ linetype     : num 1
#>   ..$ lineend      : chr "butt"
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ rect                 :List of 5
#>   ..$ fill         : chr "white"
#>   ..$ colour       : chr "white"
#>   ..$ size         : num 0.529
#>   ..$ linetype     : num 1
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ text                 :List of 11
#>   ..$ family       : chr "ArialMT"
#>   ..$ face         : chr "plain"
#>   ..$ colour       : chr "#606060"
#>   ..$ size         : num 10
#>   ..$ hjust        : num 0.5
#>   ..$ vjust        : num 0.5
#>   ..$ angle        : num 0
#>   ..$ lineheight   : num 0.9
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title           :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : num 0
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.x         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 1
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 4pt 0pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.x.top     :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 0
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 4pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.y         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : num 1
#>   ..$ vjust        : num 1.05
#>   ..$ angle        : num 0
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt -4pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.y.right   :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : num 0
#>   ..$ vjust        : num 1.05
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 0pt 4pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text            :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.x          :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "bold"
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 1
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 3.2pt 0pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.x.top      :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 0
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 3.2pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.y          :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : num 1
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 3.2pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.y.right    :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 0pt 3.2pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.ticks           : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ axis.ticks.length    : 'unit' num 4pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ axis.line            :List of 6
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ lineend      : NULL
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ axis.line.x          :List of 6
#>   ..$ colour       : chr "#606060"
#>   ..$ size         : num 0.794
#>   ..$ linetype     : NULL
#>   ..$ lineend      : chr "butt"
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ axis.line.y          : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ legend.background    :List of 5
#>   ..$ fill         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ legend.margin        : 'margin' num [1:4] 8pt 8pt 8pt 8pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ legend.spacing       : 'unit' num 16pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ legend.spacing.x     : NULL
#>  $ legend.spacing.y     : NULL
#>  $ legend.key           : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ legend.key.size      : 'unit' num 1.2lines
#>   ..- attr(*, "valid.unit")= int 3
#>   ..- attr(*, "unit")= chr "lines"
#>  $ legend.key.height    : NULL
#>  $ legend.key.width     : NULL
#>  $ legend.text          :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 10.4
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ legend.text.align    : NULL
#>  $ legend.title         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ legend.title.align   : NULL
#>  $ legend.position      : chr "right"
#>  $ legend.direction     : NULL
#>  $ legend.justification : chr "center"
#>  $ legend.box           : NULL
#>  $ legend.box.margin    : 'margin' num [1:4] 0cm 0cm 0cm 0cm
#>   ..- attr(*, "valid.unit")= int 1
#>   ..- attr(*, "unit")= chr "cm"
#>  $ legend.box.background: list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ legend.box.spacing   : 'unit' num 16pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ panel.background     : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.border         : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.spacing        : 'unit' num 8pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ panel.spacing.x      : NULL
#>  $ panel.spacing.y      : NULL
#>  $ panel.grid           :List of 6
#>   ..$ colour       : chr "#dcdfe2"
#>   ..$ size         : 'rel' num 0.667
#>   ..$ linetype     : NULL
#>   ..$ lineend      : NULL
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ panel.grid.minor     : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.grid.major.x   : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.grid.major.y   :List of 6
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ lineend      : NULL
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ panel.ontop          : logi FALSE
#>  $ plot.background      :List of 5
#>   ..$ fill         : chr "white"
#>   ..$ colour       : chr "white"
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ plot.title           :List of 11
#>   ..$ family       : chr "ArialMT"
#>   ..$ face         : chr "bold"
#>   ..$ colour       : chr "#606060"
#>   ..$ size         : num 16
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.subtitle        :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.caption         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : num 1
#>   ..$ vjust        : num 0
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.tag             :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 16
#>   ..$ hjust        : num 0.5
#>   ..$ vjust        : num 0.5
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.tag.position    : chr "topleft"
#>  $ plot.margin          : 'margin' num [1:4] 20pt 20pt 20pt 15pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ strip.background     :List of 5
#>   ..$ fill         : chr "white"
#>   ..$ colour       : chr "white"
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ strip.placement      : chr "outside"
#>  $ strip.text           :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ strip.text.x         : NULL
#>  $ strip.text.y         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : num -90
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ strip.switch.pad.grid: 'unit' num 4pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ strip.switch.pad.wrap: 'unit' num 4pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  - attr(*, "class")= chr [1:2] "theme" "gg"
#>  - attr(*, "complete")= logi TRUE
#>  - attr(*, "validate")= logi FALSE

# and this to horizontal plots, for example to horizontal bar plot
theme_thl(horizontal = TRUE)

#> List of 62
#>  $ line                 :List of 6
#>   ..$ colour       : chr "#dcdfe2"
#>   ..$ size         : num 0.529
#>   ..$ linetype     : num 1
#>   ..$ lineend      : chr "butt"
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ rect                 :List of 5
#>   ..$ fill         : chr "white"
#>   ..$ colour       : chr "white"
#>   ..$ size         : num 0.529
#>   ..$ linetype     : num 1
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ text                 :List of 11
#>   ..$ family       : chr "ArialMT"
#>   ..$ face         : chr "plain"
#>   ..$ colour       : chr "#606060"
#>   ..$ size         : num 10
#>   ..$ hjust        : num 0.5
#>   ..$ vjust        : num 0.5
#>   ..$ angle        : num 0
#>   ..$ lineheight   : num 0.9
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title           :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : num 0
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.x         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 1
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 4pt 0pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.x.top     :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 0
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 4pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.y         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : num 1
#>   ..$ vjust        : num 1.05
#>   ..$ angle        : num 0
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt -4pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.y.right   :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : num 0
#>   ..$ vjust        : num 1.05
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 0pt 4pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text            :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.x          :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 1
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 3.2pt 0pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.x.top      :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 0
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 3.2pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.y          :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "bold"
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : num 1
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 3.2pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.y.right    :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 0pt 3.2pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.ticks           : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ axis.ticks.length    : 'unit' num 4pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ axis.line            :List of 6
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ lineend      : NULL
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ axis.line.x          : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ axis.line.y          :List of 6
#>   ..$ colour       : chr "#606060"
#>   ..$ size         : num 0.794
#>   ..$ linetype     : NULL
#>   ..$ lineend      : chr "butt"
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ legend.background    :List of 5
#>   ..$ fill         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ legend.margin        : 'margin' num [1:4] 8pt 8pt 8pt 8pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ legend.spacing       : 'unit' num 16pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ legend.spacing.x     : NULL
#>  $ legend.spacing.y     : NULL
#>  $ legend.key           : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ legend.key.size      : 'unit' num 1.2lines
#>   ..- attr(*, "valid.unit")= int 3
#>   ..- attr(*, "unit")= chr "lines"
#>  $ legend.key.height    : NULL
#>  $ legend.key.width     : NULL
#>  $ legend.text          :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 10.4
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ legend.text.align    : NULL
#>  $ legend.title         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ legend.title.align   : NULL
#>  $ legend.position      : chr "right"
#>  $ legend.direction     : NULL
#>  $ legend.justification : chr "center"
#>  $ legend.box           : NULL
#>  $ legend.box.margin    : 'margin' num [1:4] 0cm 0cm 0cm 0cm
#>   ..- attr(*, "valid.unit")= int 1
#>   ..- attr(*, "unit")= chr "cm"
#>  $ legend.box.background: list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ legend.box.spacing   : 'unit' num 16pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ panel.background     : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.border         : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.spacing        : 'unit' num 8pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ panel.spacing.x      : NULL
#>  $ panel.spacing.y      : NULL
#>  $ panel.grid           :List of 6
#>   ..$ colour       : chr "#dcdfe2"
#>   ..$ size         : 'rel' num 0.667
#>   ..$ linetype     : NULL
#>   ..$ lineend      : NULL
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ panel.grid.minor     : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.grid.major.x   :List of 6
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ lineend      : NULL
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ panel.grid.major.y   : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.ontop          : logi FALSE
#>  $ plot.background      :List of 5
#>   ..$ fill         : chr "white"
#>   ..$ colour       : chr "white"
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ plot.title           :List of 11
#>   ..$ family       : chr "ArialMT"
#>   ..$ face         : chr "bold"
#>   ..$ colour       : chr "#606060"
#>   ..$ size         : num 16
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.subtitle        :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.caption         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : num 1
#>   ..$ vjust        : num 0
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.tag             :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 16
#>   ..$ hjust        : num 0.5
#>   ..$ vjust        : num 0.5
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.tag.position    : chr "topleft"
#>  $ plot.margin          : 'margin' num [1:4] 20pt 20pt 20pt 15pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ strip.background     :List of 5
#>   ..$ fill         : chr "white"
#>   ..$ colour       : chr "white"
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ strip.placement      : chr "outside"
#>  $ strip.text           :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ strip.text.x         : NULL
#>  $ strip.text.y         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : num -90
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ strip.switch.pad.grid: 'unit' num 4pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ strip.switch.pad.wrap: 'unit' num 4pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  - attr(*, "class")= chr [1:2] "theme" "gg"
#>  - attr(*, "complete")= logi TRUE
#>  - attr(*, "validate")= logi FALSE

# axis line should be hidden if coordinate axis is not in plotting area
theme_thl(show_axis_line = FALSE)

#> List of 62
#>  $ line                 :List of 6
#>   ..$ colour       : chr "#dcdfe2"
#>   ..$ size         : num 0.529
#>   ..$ linetype     : num 1
#>   ..$ lineend      : chr "butt"
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ rect                 :List of 5
#>   ..$ fill         : chr "white"
#>   ..$ colour       : chr "white"
#>   ..$ size         : num 0.529
#>   ..$ linetype     : num 1
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ text                 :List of 11
#>   ..$ family       : chr "ArialMT"
#>   ..$ face         : chr "plain"
#>   ..$ colour       : chr "#606060"
#>   ..$ size         : num 10
#>   ..$ hjust        : num 0.5
#>   ..$ vjust        : num 0.5
#>   ..$ angle        : num 0
#>   ..$ lineheight   : num 0.9
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title           :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : num 0
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.x         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 1
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 4pt 0pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.x.top     :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 0
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 4pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.y         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : num 1
#>   ..$ vjust        : num 1.05
#>   ..$ angle        : num 0
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt -4pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.title.y.right   :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : num 0
#>   ..$ vjust        : num 1.05
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 0pt 4pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text            :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.x          :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "bold"
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 1
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 3.2pt 0pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.x.top      :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : NULL
#>   ..$ vjust        : num 0
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 3.2pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.y          :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : num 1
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 3.2pt 0pt 0pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.text.y.right    :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : chr "plain"
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : 'margin' num [1:4] 0pt 0pt 0pt 3.2pt
#>   .. ..- attr(*, "valid.unit")= int 8
#>   .. ..- attr(*, "unit")= chr "pt"
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ axis.ticks           : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ axis.ticks.length    : 'unit' num 4pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ axis.line            :List of 6
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ lineend      : NULL
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ axis.line.x          : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ axis.line.y          : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ legend.background    :List of 5
#>   ..$ fill         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ legend.margin        : 'margin' num [1:4] 8pt 8pt 8pt 8pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ legend.spacing       : 'unit' num 16pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ legend.spacing.x     : NULL
#>  $ legend.spacing.y     : NULL
#>  $ legend.key           : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ legend.key.size      : 'unit' num 1.2lines
#>   ..- attr(*, "valid.unit")= int 3
#>   ..- attr(*, "unit")= chr "lines"
#>  $ legend.key.height    : NULL
#>  $ legend.key.width     : NULL
#>  $ legend.text          :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 10.4
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ legend.text.align    : NULL
#>  $ legend.title         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ legend.title.align   : NULL
#>  $ legend.position      : chr "right"
#>  $ legend.direction     : NULL
#>  $ legend.justification : chr "center"
#>  $ legend.box           : NULL
#>  $ legend.box.margin    : 'margin' num [1:4] 0cm 0cm 0cm 0cm
#>   ..- attr(*, "valid.unit")= int 1
#>   ..- attr(*, "unit")= chr "cm"
#>  $ legend.box.background: list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ legend.box.spacing   : 'unit' num 16pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ panel.background     : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.border         : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.spacing        : 'unit' num 8pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ panel.spacing.x      : NULL
#>  $ panel.spacing.y      : NULL
#>  $ panel.grid           :List of 6
#>   ..$ colour       : chr "#dcdfe2"
#>   ..$ size         : 'rel' num 0.667
#>   ..$ linetype     : NULL
#>   ..$ lineend      : NULL
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ panel.grid.minor     : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.grid.major.x   : list()
#>   ..- attr(*, "class")= chr [1:2] "element_blank" "element"
#>  $ panel.grid.major.y   :List of 6
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ lineend      : NULL
#>   ..$ arrow        : logi FALSE
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_line" "element"
#>  $ panel.ontop          : logi FALSE
#>  $ plot.background      :List of 5
#>   ..$ fill         : chr "white"
#>   ..$ colour       : chr "white"
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ plot.title           :List of 11
#>   ..$ family       : chr "ArialMT"
#>   ..$ face         : chr "bold"
#>   ..$ colour       : chr "#606060"
#>   ..$ size         : num 16
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.subtitle        :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 12
#>   ..$ hjust        : num 0
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.caption         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : num 1
#>   ..$ vjust        : num 0
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.tag             :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 16
#>   ..$ hjust        : num 0.5
#>   ..$ vjust        : num 0.5
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ plot.tag.position    : chr "topleft"
#>  $ plot.margin          : 'margin' num [1:4] 20pt 20pt 20pt 15pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ strip.background     :List of 5
#>   ..$ fill         : chr "white"
#>   ..$ colour       : chr "white"
#>   ..$ size         : NULL
#>   ..$ linetype     : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_rect" "element"
#>  $ strip.placement      : chr "outside"
#>  $ strip.text           :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : num 10
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : NULL
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ strip.text.x         : NULL
#>  $ strip.text.y         :List of 11
#>   ..$ family       : NULL
#>   ..$ face         : NULL
#>   ..$ colour       : NULL
#>   ..$ size         : NULL
#>   ..$ hjust        : NULL
#>   ..$ vjust        : NULL
#>   ..$ angle        : num -90
#>   ..$ lineheight   : NULL
#>   ..$ margin       : NULL
#>   ..$ debug        : NULL
#>   ..$ inherit.blank: logi TRUE
#>   ..- attr(*, "class")= chr [1:2] "element_text" "element"
#>  $ strip.switch.pad.grid: 'unit' num 4pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  $ strip.switch.pad.wrap: 'unit' num 4pt
#>   ..- attr(*, "valid.unit")= int 8
#>   ..- attr(*, "unit")= chr "pt"
#>  - attr(*, "class")= chr [1:2] "theme" "gg"
#>  - attr(*, "complete")= logi TRUE
#>  - attr(*, "validate")= logi FALSE